backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
plaid-python
cryptography
